+++ processor/src/test/resources/input/com/example/observe/TrackedOnDepsChangeAcceptsObserverModel.java
@@ -1,0 +1,22 @@
+package com.example.observe;
+
+import arez.Observer;
+import arez.annotations.ArezComponent;
+import arez.annotations.Executor;
+import arez.annotations.Observe;
+import arez.annotations.OnDepsChange;
+import javax.annotation.Nonnull;
+
+@ArezComponent
+public abstract class TrackedOnDepsChangeAcceptsObserverModel
+{
+  @Observe( executor = Executor.EXTERNAL )
+  public void render( final long time, float someOtherParameter )
+  {
+  }
+
+  @OnDepsChange
+  final void onRenderDepsChange( @Nonnull final Observer observer )
+  {
+  }
+}